Separate flashing into it's own file
